[mod] int to long on C_1 ABC313

+++ BeginnerContest/313/_1/ApproximateEqualization2.java
@@ -5,7 +5,7 @@ public class ApproximateEqualization2 {
 
         IOHandler io = new IOHandler();
         int n = io.nextInt();
-        int[] a = io.nextIntArray(n);
+        long[] a = io.nextLongArray(n);
         io.close();
 
         // ポイント：不変量に注目する
@@ -17,14 +17,14 @@ public class ApproximateEqualization2 {
         // A に対して操作を行った後の数列を B とする(A は操作前にソートされているとする)
         // A1, A2, ..., An-1, An, B1, B2, ..., Bn-1, Bn に対して、
         // |Ai-Bi| がなるべく小さくなるようにしたい (逆に、最小でない場合は、無駄に操作している事になる)
-        // 1回の操作で、Σ|Ai-Bi| は ±2 変動するので、Σ|Ai-Bi|/2 が求める操作回数
+        // 1回の操作で、Σ|Ai-Bi| は +2 変動するので、Σ|Ai-Bi|/2 が求める操作回数
         // さらに、その中でも、操作回数が最小になるものを求めればよい
 
         long sum = Arrays.stream(a).sum();
         Arrays.sort(a);
-        int[] b = new int[n];
-        Arrays.fill(b, (int)sum/n);
-        for (int i = 0; i < sum % n; i++) b[n - 1 - i]++;
+        long[] b = new long[n];
+        Arrays.fill(b, sum/n);
+        for (int i = 0; i < sum%n; i++) b[n - 1 - i]++;
 
         long result = 0;
         for (int i = 0; i < n; i++) result += Math.abs(a[i] - b[i]);
@@ -35,9 +35,9 @@ public class ApproximateEqualization2 {
         private Scanner sc = new Scanner(System.in);
         private void close() {this.sc.close();}
         private int nextInt() {return this.sc.nextInt();}
-        private int[] nextIntArray(int size) {
-            int[] array = new int[size];
-            for (int i = 0; i < size; i++) array[i] = this.sc.nextInt();
+        private long[] nextLongArray(int size) {
+            long[] array = new long[size];
+            for (int i = 0; i < size; i++) array[i] = this.sc.nextLong();
             return array;
         }
         private void output(long result) {System.out.println(result);}
